fix: harden api security and data integrity
- Resolve TypeScript decoration errors by augmenting FastifyInstance for RBAC
- Synchronize Drizzle schemas and update package types for idempotency, tax, and fees
- Implement restaurant ownership middleware to enforce secure resource access
- Standardize authentication flow and user profile mapping across all API routes
- Update Zod validation logic from deprecated .errors to .issues (Zod v3)
- Correct numeric type mismatches in database seeding and insert logic
- Fix missing ApiResponse imports and ensure strict type safety in order handling

+++ packages/db/src/schema/menu-items.ts
@@ -17,7 +17,7 @@ export const menuItems = pgTable("menu_items", {
     .references(() => restaurants.id),
   name: jsonb("name").notNull(),
   description: jsonb("description"),
-  price: numeric("price").notNull(),
+  price: integer("price").notNull(),
   category: text("category"),
   imageUrl: text("image_url"),
   isAvailable: boolean("is_available").default(true),
